Added taxPlayer() to Game which taxes a player when they land on a property owned by another player.
Updated incrementBalance() and decrementBalance() in Player to accept float as a parameter since the parameter being passed won't always be an int.

+++ src/Model/Player.java
@@ -40,11 +40,11 @@ public class Player {
         return position;
     }
 
-    public void decrementBalance(int amount) {
+    public void decrementBalance(float amount) {
         this.balance -= amount;
     }
 
-    public void incrementBalance(int amount) {
+    public void incrementBalance(float amount) {
         this.balance += amount;
     }
